Setting icon on Linux, Unix, macOS, Solaris throws exception
+ Added Comments to logger
+ Removed Unnecessary functions in logger
+ Made setting icon throw exception on Linux, Unix, Solaris
+ Made it so that it also prints a SystemCompatibilityError (logger)
+ Deleted @Deprecated DemoPlatformer.java
